[BUGFIX] `SceneLoader.ImportMesh` now passes error handler to `Tools.LoadFile`
=======
The Bug
=======
If `Tools.LoadFile` fails to find the resource, it will throw an error if the
optional error handler is falsy.

`SceneLoader.ImportMesh` takes an error handler, but fails to pass it to
`Tools.LoadFile`, meaning that if it attempts to load a file that
doesn't exist, `Tools.LoadFile` will throw an error that is not handled
by `SceneLoader.ImportMesh`'s `onerror` callback function.

This leads to odd behavior when attempting to Import a mesh, passing it
an `onerror` handler, and having the function throw an error that isn't
passed to the `onerror` handler.

=======
The Fix
=======
Pipe `SceneLoader.ImportMesh`'s `onerror` callback through to `Tools.LoadFile`.